itest: move channel graph related tests into one file
